#51 Refactor "ProfileCard"
